loads two books statically. Need to fix book icons
