+++ app/src/main/java/com/example/android/sqliteweather/data/BooksResponse.java
@@ -4,6 +4,5 @@ import java.io.Serializable;
 
 public class BooksResponse implements Serializable {
     public Item[] items;
-
 }
