+++ app/src/main/java/com/example/android/sqliteweather/data/LoadBookTask.java
@@ -1,13 +1,11 @@
 package com.example.android.sqliteweather.data;
 
 import android.os.AsyncTask;
-import android.util.Log;
 
 import com.example.android.sqliteweather.utils.NetworkUtils;
-import com.example.android.sqliteweather.utils.OpenWeatherMapUtils;
+import com.example.android.sqliteweather.utils.GoogleBooksUtils;
 
 import java.io.IOException;
-import java.util.ArrayList;
 
 /*
  * This is our AsyncTask for loading forecast data from OWM.  It mirrors the AsyncTask we used
@@ -46,7 +44,7 @@ class LoadBookTask extends AsyncTask<Void, Void, String> {
     protected void onPostExecute(String s) {
         BooksResponse res = null;
         try {
-            res = OpenWeatherMapUtils.parseBookJSON(s);
+            res = GoogleBooksUtils.parseBookJSON(s);
         } catch (Exception e) {
             e.printStackTrace();
         }
